+++ CHANGELOG.md
@@ -7,8 +7,13 @@
 - Adds support for the Illuminate CTI module with Intel Profile API library
   calls and `analzyer` objects. Includes support for all API parameters and
   handles pagination automatically.
+- Adds support for Illuminate Attack Surface Intelligence including third-party
+  attack surfaces. 
 - Ability to filter all RecordList analyzer objects by a list of values using
   new `filter_in` method.
+- Ability to filter all RecordList analyzer objects by a case-insensitive
+  substring search using new `filter_substring` method. Especially useful for
+  filtering a list of Attack Surface Insights or Attack Surface Third-Party vendors.
 
 
 
@@ -20,6 +25,7 @@
 - Ensure strings are returned for firstseen / lastseen dates in certificates
   property. Was causing json encoding errors when trying to encode
   `certificates.as_dict`.
+- Add missing `duration` property to pDNS `resolutions.as_dict`